fix: move writeRecordings to OutputStream module
